Correctly handle the `topojson` field
The `topojson` field in the cartodb payloads used
for the topojson import is, sadly, a string.

This commit fixes the import process so that this
string is correctly handled. As such, some errors
may occur with invalid payloads.

If an error occurs, the import process will jump
over it and print an error message.