Add relation to location in ndc_target entity
This will allow the API to show linkage data for locations
that do not yet have a full ndc imported. This is an ugly temporary hack
and should be reverted as soon as possible!

This change requires an update to the db and a re-import of linkage data. Please
remember to:

```
rails db:migrate
rails ndc_sdg_targets:import
```