Add parent sector information to accordion
I decided to modify the accordion component in case they decide that a string format
of "parent | sector" isn't good enough and in that case we'd need to do custom styling.

An alternative could be to have a selector modify the title attribute of each sector,
in that case we wouldn't have to put the concern of checking for 'parent' attributes
inside the accordion which is something that might be a bit weird. Thoughts?