Add cleanup task to SDGs goals and targets import.
There was a stray target on my local database, as well as staging and
production, and it was going to stay there until the end of times,
because we didn't have the cleanup method on this import.

Now we have, so we should be all good.

We will need to run the sdg related tasks after deploy.