Cache user selections when building visualisations
This commit implements a feature on MyClimateWatch to ensure a better
experience to users when building visualisations.
Whenever possible the visualisation creator will try to keep the
previous selections without clearing all the dropdowns that come after
the dropdown that have been changed.

As an extra this commit also adds tooltips to some of the selectors.