Link to contained route and not directly to ndcpartnership
This was a bug introduced when we added more shared modules from the
country page.

We were wrongly linking from various Explore buttons to the
ndcpartnership pages, causing a double render of the page, given that
the links are followed from inside the iframe.

This change makes it so that the links used are for routes inside
Climate Watch, with the prefix of /contained and the flag isNdcp=true to
be used by the share buttons, though maybe that is also not needed.