Adds LTS Content overview for Summary tab and Country page
This enables the following endpoint to be used on the
LTS Country page summary tab:

* http://localhost:3000/api/v1/lts/USA/content_overview

This endpoint should also be used on the main country page for the
LTS Content Overview.

As a bonus I also added an LTS indicator URL that works just like the
NDCs indicator one, if we want to update the various pages of the LTS
Explorer to use a more semantically accurate URL:

* http://localhost:3000/api/v1/lts?location=CAN&category=mitigation&source=LTS&filter=overview

It works exactly like the existing, with the same parameters.

* http://localhost:3000/api/v1/ndcs?location=CAN&category=mitigation&source=CAIT&filter=overview